Normalize with xo6
