+++ @xen-orchestra/web-core/lib/locales/en.json
@@ -529,8 +529,6 @@
   "job:pbd-plug:pbd-connected": "One or several PBDs are already connected",
   "job:pbd-unplug:missing-pbd": "No PBDs to disconnect",
   "job:pbd-unplug:pbd-disconnected": "One or several PBDs are already disconnected",
-  "job:pif-scan:in-progress": "Scan in progress…",
-  "job:pif-scan:missing-host": "Host is required",
   "job:server-connect:missing-server": "No servers to connect",
   "job:server-disconnect:missing-server": "No servers to disconnect",
   "job:server-disconnect:server-disconnected": "One or several servers are already disconnected",
