+++ @xen-orchestra/web-core/lib/locales/fr.json
@@ -529,8 +529,6 @@
   "job:pbd-plug:pbd-connected": "Un ou plusieurs PBDs sont déjà connectés",
   "job:pbd-unplug:missing-pbd": "Aucun PBD à déconnecter",
   "job:pbd-unplug:pbd-disconnected": "Un ou plusieurs PBDs sont déjà déconnectés",
-  "job:pif-scan:in-progress": "Scan en cours…",
-  "job:pif-scan:missing-host": "L'hôte est requis",
   "job:server-connect:missing-server": "Aucun serveur à connecter",
   "job:server-disconnect:missing-server": "Aucun serveur à déconnecter",
   "job:server-disconnect:server-disconnected": "Un ou plusieurs serveurs sont déjà déconnectés",
